requirements.txt
Flask==2.2.5
opencv-python-headless==4.8.0.74
numpy==1.26.0
Pillow==10.0.0
easyocr==1.6.2
torch>=1.12.0
werkzeug==2.2.3
